# api/openapi.json
{
  "components": {
    "schemas": {
...
      "BuildRequest": {
        "properties": {
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          },
          "tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpertTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "topic": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic"
        ],
        "title": "BuildRequest",
        "type": "object"
      },
...
      "ExpertTier": {
        "enum": [
          "lite",
          "standard",
          "pro"
        ],
        "title": "ExpertTier",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SetAvatarRequest": {
        "description": "Pin an avatar, or reset to the generated default.\n\nAn explicit ``{\"avatar\": null}`` is the reset. It has to be a wrapper object\nrather than a bare nullable body so that \"reset\" and \"you forgot the body\"\nstay distinguishable.",
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpertAvatar"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "SetAvatarRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Peritus API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/experts": {
      "get": {
        "description": "The caller's workspace \u2014 their own experts only.\n\nDeliberately does not fold in the public catalog: \"my experts\" must not grow\nevery time the founder publishes something. The catalog is ``GET /catalog``.",
        "operationId": "list_experts_experts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExpertSummary"
                  },
                  "title": "Response List Experts Experts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Experts",
        "tags": [
          "experts"
        ]
      }
    },
    "/experts/build": {
      "post": {
        "description": "Enqueue a durable build job and stream its progress.\n\nThe build runs in a worker (separate process or in-process), not in this\nrequest, so it survives client disconnects and server restarts. This\nresponse tails the job's persisted event log; disconnecting no longer\ncancels the build.\n\nThe choreography \u2014 slug resolution, entitlement check, create-or-requeue,\nenqueue, hold, cancel-on-refusal \u2014 lives in\n``ExpertService.request_build``. What is left here is the mapping from its\nthree failure modes onto status codes.",
        "operationId": "build_expert_experts_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Build Expert",
        "tags": [
          "experts"
        ]
      }
    },
    "/experts/{slug}": {
      "delete": {
        "operationId": "delete_expert_experts__slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Expert",
        "tags": [
          "experts"
        ]
      },
      "get": {
        "description": "Expert detail. Readable if the caller owns it, it is public, or it is\nshared with them through a live link. ``access`` says which.",
        "operationId": "get_expert_experts__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertWithCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Expert",
        "tags": [
          "experts"
        ]
      }
    },
...
    "/experts/{slug}/avatar": {
      "put": {
        "description": "Pin this expert's picture avatar, or reset it to the generated default.\n\nOwner-scoped, like every other mutation: a published expert is readable by\neveryone and re-skinnable only by the person who built it.\n\nPUT rather than PATCH because the body replaces the whole recipe \u2014 there is\nno merge, and `{\"avatar\": null}` is the reset.",
        "operationId": "set_expert_avatar_experts__slug__avatar_put",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAvatarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertWithCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Expert Avatar",
        "tags": [
          "experts"
        ]
      }
    },
    "/experts/{slug}/build/cancel": {
      "post": {
        "description": "Cancel the active (queued or running) build for an expert.\n\nA running worker notices on its next heartbeat and aborts cooperatively; a\nqueued job simply never starts. The expert is marked failed so the UI doesn't\nshow a build that will never finish.",
        "operationId": "cancel_build_experts__slug__build_cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Build Experts  Slug  Build Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Build",
        "tags": [
          "experts"
        ]
      }
    },
    "/experts/{slug}/build/events": {
      "get": {
        "description": "Reconnect to (or re-watch) a build's progress from a cursor. Multiple clients\ncan tail the same build; pass the last `seq` you saw as `after` to resume.",
        "operationId": "build_events_experts__slug__build_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Build Events",
        "tags": [
          "experts"
        ]
      }
    },
    "/experts/{slug}/build/status": {
      "get": {
        "description": "Point-in-time job status for polling clients.",
        "operationId": "build_status_experts__slug__build_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Build Status Experts  Slug  Build Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Build Status",
        "tags": [
          "experts"
        ]
      }
    },
    "/experts/{slug}/build/usage": {
      "get": {
        "description": "What the latest build of this expert actually cost, broken down by stage.\n\nOwner-scoped: spend is not part of a catalog expert's public card.",
        "operationId": "build_usage_experts__slug__build_usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Build Usage Experts  Slug  Build Usage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Build Usage",
        "tags": [
          "experts"
        ]
      }
    },
    "/experts/{slug}/catalog": {
      "patch": {
        "description": "Curate an expert: publish/unpublish, blurb, category, tags, featured, rank.\n\nOwner-scoped: a public expert is readable by everyone and curatable only by\nthe person who built it.\n\n**Publishing and shelf order are admin-only.** The catalog is curated, and\nan owner who could set ``visibility``, ``is_featured`` or ``catalog_rank``\ncould put anything at the top of the public shelf. Owners share with a link\ninstead (``PUT /experts/{slug}/share``). Taking one's own expert *off* the\nshelf stays open to its owner \u2014 nobody should need permission to unpublish.",
        "operationId": "update_expert_catalog_experts__slug__catalog_patch",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertWithCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Expert Catalog",
        "tags": [
          "experts"
        ]
      }
    },
...
  }
}